[WIP] move "web/auth" to "storage/auth_tokens"
